Move CLI introspection tests to DB tests.
