+++ crates/tests/databases-tests/src/postgres/mod.rs
@@ -1,4 +1,6 @@
 pub mod cli_update_tests;
+pub mod cli_version3_tests;
+pub mod cli_version4_tests;
 pub mod common;
 pub mod configuration_tests;
 pub mod explain_tests;
